requirements.txt
tensorflow
tensorflow-serving-api
grpc
numpy
requests